GOVSI-1113: Use shared helpers
- Refactor the `*Helper` classes from integration tests into `shared-test`
- Remove duplicate `*Helper` classes from account management integration tests and use the `shared-test` copies instead
